requirements.txt
# Core dependencies
polars>=0.20.0
mysql-connector-python>=8.0.0
jinja2>=3.1.0

# ML and data processing
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.65.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Environment management
python-dotenv>=1.0.0
